Store OtrEngineHost instead of fragmenter instructions.
Store host instance instead of instructions. Query host for
fragmentation instructions just before we start computation of numbers
or fragmentation of message, such that we can optimally adjust to
current state of host's transport.
This also means that instructions may vary for each use of the
fragmenter, e.g. each message that needs to be fragmented.