Fix bad logic w.r.t. starting new OTR session.
At the start of an OTR session, we check the current policy for
requirements w.r.t. the session. The policy should read that if both
OTRv2 and OTRv3 are disallowed, we throw an
UnsupportedOperationException. Instead, we throw an
UnsupportedOperationException as soon as one of the two is disallowed.
This is now fixed.